Bluetooth: Controller: Fix connection update window offset data type
Fix regression in refactored LLCP using uint16_t instead of
uint32_t in storing the win_offset_us value. This caused
connection update to fail with incorrect window offset being
used to schedule the connection radio events.

Regression since commit e1c2c36f569a ("Bluetooth: controller:
llcp: set refactored as default").

(cherry picked from commit 2579690be7ecf6d19bea9df74a39ecb4c924bfc8)

+++ subsys/bluetooth/controller/ll_sw/ull_conn_internal.h
@@ -81,7 +81,7 @@ static inline void cpr_active_reset(void)
 uint16_t ull_conn_event_counter(struct ll_conn *conn);
 
 void ull_conn_update_parameters(struct ll_conn *conn, uint8_t is_cu_proc,
-				uint8_t win_size, uint16_t win_offset_us,
+				uint8_t win_size, uint32_t win_offset_us,
 				uint16_t interval, uint16_t latency,
 				uint16_t timeout, uint16_t instant);
